refactor: 구간DTO 패키지 변경 (이동)
